Caching! And other stuff
It took a while but I finally implemented data caching. This way the data is only loaded once on start and saved every 10 minutes (and on server stop). This will prevent future lag when the data files become super large. This only applies to data files since config files are negligible.

I also tweaked up a ton of other code but I did it along the way so I don't remember a lot.
- Axe will now have it's durability changed before returning it so that it doesn't have a weird double animation.
- Moved all the placing stuff to PlaceManager so that LockManager isn't so big
- Renamed ConfigManager to FileManager to better fit what it does
- Added more JavaDocs